+++ contracts/libraries/LibERC998.sol
@@ -37,7 +37,7 @@ library ERC998 {
     InstallationAppStorage storage s = LibAppStorageInstallation.diamondStorage();
     s.nftInstallationBalances[_toContract][_toTokenId][_id] += _value;
     if (s.nftInstallationIndexes[_toContract][_toTokenId][_id] == 0) {
-      s.nftInstallations[_toContract][_toTokenId].push(uint16(_id));
+      s.nftInstallations[_toContract][_toTokenId].push(_id);
       s.nftInstallationIndexes[_toContract][_toTokenId][_id] = s.nftInstallations[_toContract][_toTokenId].length;
     }
   }
@@ -50,7 +50,7 @@ library ERC998 {
     InstallationAppStorage storage s = LibAppStorageInstallation.diamondStorage();
     s.ownerInstallationBalances[_to][_id] += _value;
     if (s.ownerInstallationIndexes[_to][_id] == 0) {
-      s.ownerInstallations[_to].push(uint16(_id));
+      s.ownerInstallations[_to].push(_id);
       s.ownerInstallationIndexes[_to][_id] = s.ownerInstallations[_to].length;
     }
   }
@@ -70,7 +70,7 @@ library ERC998 {
       uint256 lastIndex = s.ownerInstallations[_from].length - 1;
       if (index != lastIndex) {
         uint256 lastId = s.ownerInstallations[_from][lastIndex];
-        s.ownerInstallations[_from][index] = uint16(lastId);
+        s.ownerInstallations[_from][index] = lastId;
         s.ownerInstallationIndexes[_from][lastId] = index + 1;
       }
       s.ownerInstallations[_from].pop();
@@ -94,7 +94,7 @@ library ERC998 {
       uint256 lastIndex = s.nftInstallations[_fromContract][_fromTokenId].length - 1;
       if (index != lastIndex) {
         uint256 lastId = s.nftInstallations[_fromContract][_fromTokenId][lastIndex];
-        s.nftInstallations[_fromContract][_fromTokenId][index] = uint16(lastId);
+        s.nftInstallations[_fromContract][_fromTokenId][index] = lastId;
         s.nftInstallationIndexes[_fromContract][_fromTokenId][lastId] = index + 1;
       }
       s.nftInstallations[_fromContract][_fromTokenId].pop();
